% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpop-package.R
\docType{package}
\name{gpop-package}
\alias{gpop}
\alias{gpop-package}
\title{Population Genetics Models and Analysis}
\description{
Initial development foundations. Scientific interfaces are proposals only;
no scientific functions are exported. Base-R reference arithmetic is
installed under examples, separately from the future method interfaces.
}
\author{
\strong{Maintainer}: Peter Soerensen \email{peter.sorensen@r-qgg.org}

}
\keyword{internal}
